Split fs scaning and metadata matching
