+++ scanner/.dockerignore
@@ -1,0 +1,2 @@
+Dockerfile*
+
