+++ scanner/monitor/__main__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+
+import asyncio
+import monitor
+
+asyncio.run(monitor.main())
