+++ scanner/monitor/requirements.txt
@@ -1,0 +1,1 @@
+aio-pika
